third_party/blink/web_tests/external/wpt/webtransport/sendorder.https.any.worker-expected.txt

This is a testharness.js-based test.
[FAIL] WebTransport client should be able to create and handle a bidirectional stream with sendOrder
  assert_equals: expected 3 but got 0
[PASS] WebTransport client should be able to modify default sendOrder after stream creation
[FAIL] WebTransport client should be able to modify existing sendOrder after stream creation
  assert_equals: expected 3 but got 0
Harness: the test ran to completion.
